BGDIINF_SB-3071: Moved the external-layer-providers.json in its module
Move it into its module as well as its utilities, which are highly related to
the import tool.

Also avoid a typescript failure in vscode by adding explicitly json file in
the tsconfig.json include pattern.
