+++ src/modules/infobox/utils/external-layer-providers.json
