+++ src/modules/infobox/utils/external-provider.js
